Reduce button clicking boilerplate
Update the make_button function to return a dictionary of index,
label, and rect. Create buttons by iterating through a list of config
parameters, and storing the output in a list. When a click occurs,
iterate this list of buttons looking for the first rect which returns
true for the pygame.Rect.collidepoint() method. Handle any non-button
clicks by logging the event and moving on. When we do have a hit,
log the click, and use the matched button's index property to call
button_handler.